requirements.txt
torch==2.2.1
torchaudio==2.2.1
hydra-core==1.3.1
einops==0.4.1
pandas==2.2.1
soundfile==0.12.1
tqdm==4.66.3
POT==0.9.3
